app/requirements.txt
langchain
langchain-groq
langchain-openai
gradio
gradio_toggle
